+++ src/auth/strategies/jwt-optional.strategy.ts
